Version del chatbot, procesa y contesta mensajes
