+++ package.json
@@ -10,13 +10,16 @@
     "preview": "vite preview",
     "start": "vite --port 3000",
     "deploy": "gh-pages -d dist",
-    "homepage": "/BreinLogic/"
+    "homepage": "/BreinLogic"
   },
   "dependencies": {
     "axios": "^1.8.4",
     "dotenv": "^16.5.0",
+    "file-saver": "^2.0.5",
     "lucide-react": "^0.344.0",
     "node": "^18.20.7",
+    "pdf-lib": "^1.17.1",
+    "pdf-parse": "^1.1.1",
     "pdfjs-dist": "^5.1.91",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
