+++ src/components/EmptyChat.jsx
@@ -1,0 +1,8 @@
+export default function EmptyChat() {
+  return (
+    <div className="h-full flex flex-col items-center justify-center">
+      <div className="text-4xl font-bold text-blue-500 mb-4">BreinLogic</div>
+      <div className="text-xl text-gray-400">How can I help you today?</div>
+    </div>
+  );
+}
