+++ vite.config.js
@@ -5,5 +5,5 @@ import react from "@vitejs/plugin-react";
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
-  base: "/BreinLogic/",
+  base: "/BreinLogic",
 });
